System Updated for Perfomance

+++ Engine/Updater.bash
@@ -1,14 +1,16 @@
-RootFolder=Engine/Core
-Containers=$RootFolder/Containers
-SambaFolder=$RootFolder/Samba
-Includes=$RootFolder/Includes
-SambaFile=/etc/samba/smb.conf
+# Make Sure Original File Exisits
+if [ ! -f /Engine/Sources.bash ]; then
+    wget "https://raw.githubusercontent.com/EngineFw/DockerShell/main/Engine/Sources.bash" -O "/Engine/Sources.bash"
+else
+    if [ ! -d "/Engine/NV" ]; then
+        mkdir -p /Engine/NV
+    fi
+    wget "https://raw.githubusercontent.com/EngineFw/DockerShell/main/Engine/Sources.bash" -O "/Engine/NV/Sources.bash"
 
-# Download the newest Version of sources
-if [ ! -d /Engine/Core/NV ]; then
-    mkdir -p /Engine/Core/NV
 fi
 
-wget "https://raw.githubusercontent.com/EngineFw/DockerShell/main/Engine/Sources.bash" -O "/Engine/NV/Sources.bash"
-
 # grep 'version=' ./Sources.bash
+
+DownloadNV() {
+
+}
